feat: --realtime-latencies live per-second percentile output
Replaces the periodic '[RUN #N ...] ops/sec' progress line — which only
exposed averages — with a per-tick block that surfaces miss ratio and the
percentiles configured by --print-percentiles. Default behavior (no flag)
is byte-identical to before, so existing log scrapers keep working.

Example (--print-percentiles=50,99,99.9,99.99,99.99999,99.999999):

    [RUN #1  33%   1s] throughput 153,287 (avg: 153,287) ops/sec   8.62MB/sec (avg: 8.62MB/sec)   miss 49.86% (avg: 49.86%)
    [RUN #1  33%   1s] latency  p50  1.215 (avg:  1.215)  p99  2.415 (avg:  2.415)  p99.9  9.151 (avg:  9.151)
    [RUN #1  33%   1s] latency  p99.99 16.895 (avg: 16.895)  p99.99999 18.559 (avg: 18.559)  p99.999999 18.559 (avg: 18.559) ms

* Each metric carries both an immediate value and an '(avg: X)' overall
  counterpart. Overall percentiles come from a run-lifetime histogram we
  hdr_add into across ticks (allocated once, freed after the loop).
* Throughput line: ops/sec (cur/avg), bytes/sec (cur/avg), miss ratio
  (per-second / cumulative). The legacy absolute ops counter is dropped
  under --realtime-latencies — it scrolled fast and meant little.
* Latency line wraps at 3 percentiles per line so a 6-percentile config
  doesn't blow out the terminal. Line count is deterministic given the
  configured list, so the in-place TTY redraw stays stable across ticks.
* On a TTY the block redraws in place via \033[NA + \033[K. On non-TTY
  (CI logs, redirected files) it just appends, producing a clean
  per-second time series each line of which can be parsed independently
  (every line carries the [RUN ... Ts] tag).

Supporting fixes pulled in along the way:

* Percentile labels: replace p%.*f / p%g with p%.10g everywhere a label
  is rendered (final stats table, --realtime-latencies output, statsd
  metric names). The previous formatter capped at 3 fractional digits
  and silently rounded 99.99999 and 99.999999 both to 'p100', colliding
  in the table output. JSON keys keep their p%.2f / p%.3f shape for
  back-compat with existing consumers and the test suite.
* Promote config_quantiles::quantile_list from vector<float> to
  vector<double> (parse via strtod). Avoids deep-tail precision loss at
  storage time — float couldn't represent 99.999999 distinctly from 100.
* Auto-size percentile latency columns in the final stats table so long
  headers like 'p99.999999 Latency' don't spill into the adjacent column
  and break the table separators. Standard p50/p99/p99.9 keep width 15.

run_stats / client_group:

* Expose get_total_hits() / get_total_misses() so the periodic loop can
  read per-second and cumulative cache stats without a full merge.

Tests:

* tests/test_realtime_latencies.py covers (1) the 2-line block shape with
  immediate + (avg: X) for ops/sec, bytes/sec, miss%, and each configured
  percentile, (2) the 6-percentile wrap path, (3) deep-tail p99.99999 /
  p99.999999 labels not rounding to 'p100' (both in the stderr stream and
  the final stats table header), and (4) a regression test that the
  default progress line (no flag) is byte-identical to before.

Man page and bash completion updated.

+++ config_types.h
@@ -52,12 +52,12 @@ struct config_ratio
 
 struct config_quantiles
 {
-    std::vector<float> quantile_list;
+    std::vector<double> quantile_list;
     config_quantiles();
     config_quantiles(const char *ratio_str);
     bool is_defined(void);
-    inline std::vector<float>::iterator begin() { return quantile_list.begin(); }
-    inline std::vector<float>::iterator end() { return quantile_list.end(); }
+    inline std::vector<double>::iterator begin() { return quantile_list.begin(); }
+    inline std::vector<double>::iterator end() { return quantile_list.end(); }
 };
 
 struct config_weight_list
